+++ FutsalAgentSimulator/src/pt/ulisboa/aasma/fas/j2d/Sprite.java
@@ -4,6 +4,6 @@ import java.awt.Graphics2D;
 
 
 public interface Sprite {
-	void update(long time);
+	void update(double time);
 	void draw(Graphics2D g2d);
 }
